Switched from linear interpolation of tauh to taking the 3D field as ground truth (so blocky cells of homogeneous characteristics)

+++ include/util.h
@@ -26,14 +26,14 @@ void LOGvec(const std::vector<T>& arr, const std::string& name = "Array", bool e
     }
 }
 
-inline float kahan_sum(const std::vector<float>& values) 
+inline double kahan_sum(const std::vector<double>& values) 
 {
     // from Mr. Chat
-    float sum = 0.0;
-    float c = 0.0;  // Compensation
-    for (float x : values) {
-        float y = x - c;
-        float t = sum + y;
+    double sum = 0.0;
+    double c = 0.0;  // Compensation
+    for (double x : values) {
+        double y = x - c;
+        double t = sum + y;
         c = (t - sum) - y;
         sum = t;
     }
@@ -43,7 +43,7 @@ inline float kahan_sum(const std::vector<float>& values)
 
 namespace cf
 {
-    const float PI = 3.1415927;
-    const float RHO = 1.225;
-    const float CP = 1004;
+    const double PI = 3.1415927;
+    const double RHO = 1.225;
+    const double CP = 1004;
 }
